Deleted the Read comp from admin dash
